requirements.txt
torch
ultralytics
easyocr
opencv-python-headless  
streamlit
Pillow
numpy
